一旦履歴でました！

+++ demo/src/main/java/com/example/demo/repository/QuizRecordRepository.java
@@ -6,9 +6,12 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.QuizRecord;
 
 @Repository
-public interface QuizRecordRepository extends JpaRepository<QuizRecord, Integer> {
+public interface QuizRecordRepository extends JpaRepository<QuizRecord, Long> {
     List<QuizRecord> findAllByOrderByCreatedAtDesc(); // 最新順に表示したい
 
     List<QuizRecord> findByUserId(Long userId);
     List<QuizRecord> findByChapter(int chapter);
+    // QuizRecordRepository.java
+    
+
 }
